+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.0] - ?
 
+- add admin command to toggle debug logging
 - rename project
 - update code to match csharp standards
 - wrap debug logging in disabled debug flag